requirements.txt
flask
flask_bcrypt
flask_login
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
flask_migrate
flask_wtf
wtforms
Pillow
python-dotenv
email_validator
psycopg2
